pull out FoldNull tests
